ILogger on steroids!
ILogger can now do nice stuff like
  m_logger.debug() << "Yipee!";

The SHAREMIND_DEFINE_LOGS and SHAREMIND_DEFINE_PREFIXED_LOG will in the future
be deprecated in favor of this new behaviour. Prefixed wrappers (and wrappers
of wrapper etc...) can now be defined like this:

  template <class Logger> // Might be ILogger or another wrapper
  struct MyClass {
    MyClass(Logger & logger)
      : m_logger(logger.wrap("[MyClass] "))
    {
      m_logger.info() << "Yipee!";
    }
    typename Logger::Wrapped m_logger;
  };

Note however that this behaviour comes with a cost in C++03 and earlier. Namely
the object the functions fatal(), error(), warning(), info(), debug() and
fullDebug() return are not "copyable" and "copy-assignable" in C++03 and
earlier. Both of these semantics are actually used to "move" these objects as
done in C++11 and later. Fortunately copy-semantics are not usually meant to be
used by users of the ILogger (and wrappers) anyway, but if you do, know that

   Y O U   H A V E   B E E N   W A R N E D  !

Anyway this helped me to make the networking library not depend on the Common
library.