Attempt a workaround thread-local storage (TLS) not being available in Clang.
TLS is Clang is only available when using Clang with GNU libstdc++ 4.8 or later.
However there's no good way to determine the version of libstdc++ using
preprocessor macros. The __GLIBCXX__ macro gives us a date, but unfortunately
its value in libstdc++ 4.7.4 is greater than in 4.8.0, hence it is not entirely
reliable.